-Sacar la interface a un archivo a parte e importarlo donde es necesario -Crear componente para editar las tareas
