+++ src/Task.ts
@@ -1,0 +1,9 @@
+export interface Task {
+    id: string;
+    title: string;
+    description: string;
+    dateCreated: string;
+    deadLine: string;
+    priority: string;
+    doing: boolean;
+}
