+++ src/components/ListItem.tsx
@@ -1,4 +1,6 @@
+import { useState } from "react";
 import { Pencil, Trash } from "./Icons";
+import  ModalEdit  from './ModalEdit';
 
 interface ItemProps {
   title: string;
@@ -6,9 +8,7 @@ interface ItemProps {
   onDelete: () => void;
 }
 
-
-
-export function ListItem({ title, description, onDelete}: ItemProps) {
+export function ListItem({ title, description, onDelete }: ItemProps) {
   const styles = {
     deleteButton: {
       color: "#9d0208",
@@ -17,18 +17,30 @@ export function ListItem({ title, description, onDelete}: ItemProps) {
       color: "#f5bd1f",
     },
   };
+
+  const [show, setShow] = useState(false);
+
+  const showModal = () => setShow(true);
+  const hideModal = () => setShow(false);
+
+
   return (
-    <div className="list-item">
-      <h1>{title}</h1>
-      {description && <p>{description}</p>}
-      <div>
-        <button style={styles.editButton}>
-          <Pencil />
-        </button>
-        <button style={styles.deleteButton} onClick={onDelete}>
-          <Trash />
-        </button>
+    <>
+      {/* <ModalEdit show={show} hideModal={hideModal} onSubmit={handleEditTask} /> */}
+      <div className="list-item"
+        onClick={showModal}
+      >
+        <h1>{title}</h1>
+        {description && <p>{description}</p>}
+        <div>
+          <button style={styles.editButton}>
+            <Pencil />
+          </button>
+          <button style={styles.deleteButton} onClick={onDelete}>
+            <Trash />
+          </button>
+        </div>
       </div>
-    </div>
+    </>
   );
 }
